Add files via upload
This file is intended to provide guidance for BCI developers that are new to the field and who would like some developmental support in the form of working code. Please cite the following research if code is used for research:

Harnarinesingh, R.E.S. and C.S. Syan (2018), Investigation of the Mirrored-Word Reading Paradigm for BCI Implementation, Biomedical Engineering / Biomedizinische Technik, vol. 36, Ahead of Print